Move TOML parse and dump implementations from PyprojectTOMLManager to TOMLFileManager Add tests for TOMLFileManager
